refactor(desktop): split large React components into focused sections
Extract SettingsPage (740→104 lines) into 7 section components under
pages/settings/: OutputSection, FormatDefaultsSection, MediaSection,
SearchSection, CookiesSection, NetworkSection, AudioNormSection.

Extract SearchPage callbacks into useSearchActions hook (454→364 lines),
encapsulating download, format dialog, and search navigation actions.

All 281 Vitest tests pass unchanged.
