+++ crates/rdlp-desktop/src/pages/settings/types.ts
@@ -1,0 +1,7 @@
+import type { AppSettings } from "../../types";
+
+/** Props shared by all settings section components. */
+export interface SettingsSectionProps {
+    draft: AppSettings;
+    onChange: (next: AppSettings) => void;
+}
